story #8451: Replace the global save button with a local save button on Kanban label

+++ plugins/agiledashboard/www/js/kanban/src/app/edit-kanban/edit-kanban-controller.js
@@ -2,16 +2,18 @@ angular
     .module('kanban')
     .controller('EditKanbanCtrl', EditKanbanCtrl);
 
-EditKanbanCtrl.$inject = ['$scope', '$window', '$modalInstance', 'KanbanService', 'kanban', 'augmentColumn', 'SharedPropertiesService', 'gettextCatalog'];
+EditKanbanCtrl.$inject = ['$scope', '$window', '$modalInstance', 'KanbanService', 'kanban', 'augmentColumn', 'updateKanbanName', 'SharedPropertiesService', 'gettextCatalog'];
 
-function EditKanbanCtrl($scope, $window, $modalInstance, KanbanService, kanban, augmentColumn, SharedPropertiesService, gettextCatalog) {
+function EditKanbanCtrl($scope, $window, $modalInstance, KanbanService, kanban, augmentColumn, updateKanbanName, SharedPropertiesService, gettextCatalog) {
     var self = this;
 
     _.extend(self, {
         kanban                   : kanban,
         saving                   : false,
-        cancel                   : cancel,
         deleting                 : false,
+        confirm_delete           : false,
+        saving_new_column        : false,
+        cancel                   : cancel,
         adding_column            : false,
         new_column_label         : '',
         reorderColumnsTreeOptions: {
@@ -27,7 +29,9 @@ function EditKanbanCtrl($scope, $window, $modalInstance, KanbanService, kanban, 
     function saveModifications() {
         self.saving = true;
         KanbanService.updateKanbanLabel(kanban.id, kanban.label).then(function () {
-            $modalInstance.close(kanban);
+            self.saving = false;
+            updateKanbanName(kanban);
+
         }, function (response) {
             $modalInstance.dismiss(response);
         });
@@ -38,22 +42,27 @@ function EditKanbanCtrl($scope, $window, $modalInstance, KanbanService, kanban, 
     }
 
     function deleteKanban() {
-        self.deleting = true;
+        if (self.confirm_delete) {
+            self.deleting = true;
 
-        KanbanService.deleteKanban(kanban.id).then(function () {
-            var message = gettextCatalog.getString(
-                'Kanban {{ label }} successfuly deleted',
-                { label: kanban.label }
-            );
-            $window.sessionStorage.setItem('tuleap_feedback', message);
-            $window.location.href = '/plugins/agiledashboard/?group_id=' + SharedPropertiesService.getProjectId();
-        }, function (response) {
-            $modalInstance.dismiss(response);
-        });
+            KanbanService.deleteKanban(kanban.id).then(function () {
+                var message = gettextCatalog.getString(
+                    'Kanban {{ label }} successfuly deleted',
+                    { label: kanban.label }
+                );
+                $window.sessionStorage.setItem('tuleap_feedback', message);
+                $window.location.href = '/plugins/agiledashboard/?group_id=' + SharedPropertiesService.getProjectId();
+            }, function (response) {
+                $modalInstance.dismiss(response);
+            });
+
+        } else {
+            self.confirm_delete = true;
+        }
     }
 
     function processing() {
-        return self.deleting || self.saving;
+        return self.deleting || self.saving || self.saving_new_column;
     }
 
     function cancelAddColumn() {
@@ -62,6 +71,8 @@ function EditKanbanCtrl($scope, $window, $modalInstance, KanbanService, kanban, 
 
     function addColumn() {
         if (self.adding_column) {
+            self.saving_new_column = true;
+
             KanbanService.addColumn(kanban.id, self.new_column_label).then(function(column_representation) {
                 var new_column = column_representation.data;
 
@@ -71,8 +82,9 @@ function EditKanbanCtrl($scope, $window, $modalInstance, KanbanService, kanban, 
 
                 kanban.columns.push(new_column);
 
-                self.adding_column    = false;
-                self.new_column_label = '';
+                self.adding_column     = false;
+                self.saving_new_column = false;
+                self.new_column_label  = '';
 
             }, function (response) {
                 $modalInstance.dismiss(response);
